Aggiunto il download da PokeAPI e il salvataggio dei dati su Room

+++ app/src/main/java/it/codeclub/pokeclub/db/entities/Ability.kt
@@ -5,10 +5,11 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class Ability (
-    @PrimaryKey val abilityId: Int,
+    @PrimaryKey val abilityId: Long,
     val name_it: String,
     val effect_it: String,
     val name_en: String,
     val effect_en: String,
     val isSpecial: Boolean = false
+
 )
